Refactor: Cleanup essentials. (#1383)
- Refactor main Essentials' components
- Highlight searched words, include more columns for search
- Remove empty file and console log statements
- Fix issue around infinite scroll and searching
- Remove commented console logs
- Fix number hyperlinking issue
- Use feather icons
- Add loader gif for infinite scroll
- Change stick thing
- Revert width for search bar
- Change margins for mobile view
- Change margins to fix desktop view
- Support category sorting and Google Tag tracker (#1304)
- Change pages to page in resourctable.js
- Sort the categories
- Tag code snippet for Google Tag management
