+++ package.json
@@ -6,7 +6,6 @@
   "dependencies": {
     "@juggle/resize-observer": "^3.1.3",
     "@material-ui/core": "^4.9.9",
-    "@material-ui/icons": "^4.9.1",
     "@testing-library/jest-dom": "^5.3.0",
     "@testing-library/react": "^10.0.1",
     "animejs": "^3.2.0",
