+++ src/App.scss
@@ -532,8 +532,8 @@ h6 {
   flex-direction: row;
   justify-content: center;
   flex-wrap: wrap;
-  margin-left: 7rem;
-  margin-right: 5rem;
+  margin-left: 5.5rem;
+  margin-right: 5.5rem;
 
   .resource-left,
   .resource-right {
@@ -552,10 +552,11 @@ h6 {
     margin-right: 2.5rem;
   }
 
-  .tableandcontrols{
+  .tableandaccordions{
     display: flex;
     flex-direction: column;
     width: 100%;
+    align-items: center;
   }
 
   table {
@@ -566,6 +567,9 @@ h6 {
     text-transform: none;
     border-spacing: 3px 2px;
     border-collapse: separate;
+    table-layout: fixed;
+    // animation-name: fadeInOut;
+    // animation-duration: 2s;
 
     thead {
       background: $purple-light;
@@ -579,9 +583,12 @@ h6 {
         border-radius: 5px;
         transition: all 0.1s ease-in-out;
         z-index: 99;
+        &.descriptionCol{
+          width: 50%;
+        }
 
         &.sticky {
-          top: 4px;
+          top: 0;
           background: $gray-opaque;
         }
 
@@ -811,8 +818,8 @@ h6 {
   font-size: 0.7rem !important;
   font-weight: 400;
   text-align: center;
-  margin-top: 0.2rem;
-  margin-bottom: 0.1rem;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
   width: 100%;
 
   .react-autosuggest__input {
